middleware stuff and restructuring
