+++ packages/commons/src/context/context.middleware.ts
@@ -1,5 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import { v4 as uuidv4 } from "uuid";
+import { logger } from "../logging/index.js";
 
 export const contextMiddleware = async (
   req: Request,
@@ -10,5 +11,10 @@ export const contextMiddleware = async (
     serviceName: "push",
     correlationId: uuidv4(),
   };
+  const loggerInstance = logger({
+    serviceName: req.ctx.serviceName,
+    correlationId: req.ctx.correlationId,
+  });
+  loggerInstance.info("Context init");
   next();
 };
